+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- use Django 4.1 async ORM
+
 ## [v0.1.0] - 2020-08-05
 - Initial Release